Baisc updated to get a default-type page going
